+++ src/__tests__/init-tests.ts
@@ -33,6 +33,7 @@ import { FieldNode, OperationDefinitionNode } from "graphql"
 import { httpUrl, assignDeep, cloneDeep } from "../server/utils"
 import fetch from "node-fetch"
 import { QCollectionQuery } from "../server/data/collection-query"
+import { BocStorage } from "../server/data/boc-storage"
 
 jest.setTimeout(100000)
 
@@ -82,7 +83,6 @@ export function queryText(
                 orderBy,
             },
             selectionInfo(result),
-            0,
         )?.text ?? "",
     )
 }
@@ -253,6 +253,7 @@ export function createLocalArangoTestData(logs: QLogs): QBlockchainData {
         logs: new QLogs(),
         tracer: QTracer.create(testConfig),
         stats: QStats.create("", [], 0),
+        bocStorage: new BocStorage(config.blockBocs),
         isTests: true,
         subscriptionsMode: SubscriptionsMode.Arango,
         filterConfig: config.queries.filter,
@@ -264,8 +265,6 @@ export class MockProvider<T extends QResult> implements QDataProvider {
     data: T[]
     queryCount: number
     hotUpdateCount: number
-    shards = []
-    shardingDegree = 0
 
     constructor(data: T[]) {
         this.data = data
@@ -338,6 +337,7 @@ export function createTestData(providers: QDataProviders): QBlockchainData {
         logs: new QLogs(),
         tracer: QTracer.create(testConfig),
         stats: QStats.create("", [], 0),
+        bocStorage: new BocStorage(testConfig.blockBocs),
         isTests: true,
         subscriptionsMode: SubscriptionsMode.Arango,
         filterConfig: testConfig.queries.filter,