+++ src/server/data/collection-joins.ts
@@ -7,7 +7,7 @@ import { joinFields } from "../graphql/resolvers-generated"
 import QData from "./data"
 import QBlockchainData from "./blockchain"
 import { QTraceSpan } from "../tracing"
-import { arraysAreEqual, required } from "../utils"
+import { arraysAreEqual } from "../utils"
 
 export class QJoinQuery {
     on: string
@@ -18,8 +18,6 @@ export class QJoinQuery {
     refOn: string
     refOnAttr: string
     refOnIsArray: boolean
-    shardOn?: string
-    shardOnShardingDegree: number
     canJoin: (parent: unknown, args: JoinArgs) => boolean
     fieldSelection: SelectionSetNode
 
@@ -28,7 +26,6 @@ export class QJoinQuery {
             on: string
             collection: string
             refOn: string
-            shardOn?: string
             canJoin: (parent: unknown, args: JoinArgs) => boolean
         },
         public field: FieldNode,
@@ -44,22 +41,10 @@ export class QJoinQuery {
         this.refOnIsArray = join.refOn.endsWith("[*]")
         this.refOn = this.refOnIsArray ? join.refOn.slice(0, -3) : join.refOn
         this.refOnAttr = this.refOn === "id" ? "_key" : this.refOn
-        this.shardOn = join.shardOn === "id" ? "_key" : join.shardOn
         this.canJoin = join.canJoin
         this.fieldSelection = mergeFieldWithSelectionSet(
             this.refOn,
             this.field.selectionSet,
-        )
-
-        const mainCollectionProvider = required(
-            data.collectionsByName.get(mainCollectionName)?.provider,
-        )
-        const refCollectionProvider = required(this.refCollection.provider)
-        const mainShardingDegree = mainCollectionProvider.shardingDegree
-        const refShardingDegree = refCollectionProvider.shardingDegree
-        this.shardOnShardingDegree = Math.min(
-            mainShardingDegree,
-            refShardingDegree,
         )
     }
 
@@ -86,8 +71,7 @@ export class QJoinQuery {
     }
 
     buildPlan(mainRecords: Record<string, unknown>[]): JoinPlan {
-        const plan = new Map<Shard, Map<OnValue, MainRecord[]>>()
-        const shards = new Map<OnValue, Shard>()
+        const plan = new Map<OnValue, MainRecord[]>()
         for (const record of mainRecords) {
             if (!this.canJoin(record, this.args)) {
                 continue
@@ -105,58 +89,23 @@ export class QJoinQuery {
             if (Array.isArray(onValue)) {
                 const joins: (Record<string, unknown> | null)[] = []
                 for (const value of onValue) {
-                    processValue(value, record, this)
+                    processValue(value, record)
                     joins.push(null)
                 }
                 record[this.joinField] = joins
             } else {
-                processValue(onValue, record, this)
+                processValue(onValue, record)
             }
         }
         return plan
 
-        function processValue(
-            value: string,
-            record: MainRecord,
-            joinQuery: QJoinQuery,
-        ) {
-            const shard = getShard(value, record, joinQuery)
-            let shardPlan = plan.get(shard)
-            if (shardPlan === undefined) {
-                shardPlan = new Map<OnValue, MainRecord[]>()
-                plan.set(shard, shardPlan)
-            }
-
-            const valuePlan = shardPlan.get(value)
+        function processValue(value: string, record: MainRecord) {
+            const valuePlan = plan.get(value)
             if (valuePlan !== undefined) {
                 valuePlan.push(record)
             } else {
-                shardPlan.set(value, [record])
+                plan.set(value, [record])
             }
-        }
-
-        function getShard(
-            value: string,
-            record: MainRecord,
-            joinQuery: QJoinQuery,
-        ): string | null {
-            if (joinQuery.shardOn === undefined) {
-                return null
-            }
-
-            let shard = shards.get(value)
-            if (shard !== undefined) {
-                return shard
-            }
-
-            ;[shard] = QCollectionQuery.getShards(
-                joinQuery.mainCollectionName,
-                { [joinQuery.shardOn]: { eq: record[joinQuery.shardOn] } },
-                joinQuery.shardOnShardingDegree,
-            ) ?? [null]
-
-            shards.set(value, shard)
-            return shard
         }
     }
 
@@ -202,65 +151,50 @@ export class QJoinQuery {
         const timeout = this.args.timeout ?? defaultTimeout ?? 0
         const timeLimit = Date.now() + timeout
         let joinedRecords: Record<string, unknown>[] = []
-        const shards = [...joinPlan.keys()]
-        for (const shard of shards) {
-            const shardPlan = joinPlan.get(shard)
-            if (shardPlan === undefined) {
-                throw new Error(
-                    "This error should be impossible in join fetcher",
-                )
+        let prevPortionOnValues = [] as string[]
+        while (joinPlan.size > 0) {
+            const portionOnValues = [...joinPlan.keys()].slice(0, 100)
+            if (arraysAreEqual(portionOnValues, prevPortionOnValues)) {
+                // throttle repeated queries to 5 times per second
+                await new Promise(resolve => setTimeout(resolve, 200))
             }
-            let prevPortionOnValues = [] as string[]
-            while (shardPlan.size > 0) {
-                const portionOnValues = [...shardPlan.keys()].slice(0, 100)
-                if (arraysAreEqual(portionOnValues, prevPortionOnValues)) {
-                    // throttle repeated queries to 5 times per second
-                    await new Promise(resolve => setTimeout(resolve, 200))
+            prevPortionOnValues = portionOnValues
+            const joinQuery = QCollectionQuery.createForJoin(
+                request,
+                portionOnValues,
+                this.refCollection.name,
+                this.refCollection.docType,
+                this.refOn,
+                this.refOnIsArray,
+                this.fieldSelection,
+                request.services.config,
+            )
+            if (joinQuery !== null) {
+                const fetcher = async (span: QTraceSpan) => {
+                    span.log({
+                        text: joinQuery.text,
+                        params: joinQuery.params,
+                    })
+                    return (await this.refCollection.queryProvider({
+                        text: joinQuery.text,
+                        vars: joinQuery.params,
+                        orderBy: [],
+                        isFast: true,
+                        request,
+                        traceSpan: span,
+                    })) as Record<string, unknown>[]
                 }
-                prevPortionOnValues = portionOnValues
-                const joinQuery = QCollectionQuery.createForJoin(
-                    request,
-                    portionOnValues,
-                    this.refCollection.name,
-                    this.refCollection.docType,
-                    this.refOn,
-                    this.refOnIsArray,
-                    this.fieldSelection,
-                    required(this.refCollection.provider).shardingDegree,
-                    request.services.config,
+                const joinedPortion = await traceSpan.traceChildOperation(
+                    `${mainCollection.name}.query.join`,
+                    fetcher,
                 )
-                if (joinQuery !== null) {
-                    const fetcher = async (span: QTraceSpan) => {
-                        span.log({
-                            text: joinQuery.text,
-                            params: joinQuery.params,
-                            shards: joinQuery.shards,
-                        })
-                        return (await this.refCollection.queryProvider({
-                            text: joinQuery.text,
-                            vars: joinQuery.params,
-                            orderBy: [],
-                            isFast: true,
-                            request,
-                            shards:
-                                shard !== null
-                                    ? new Set([shard])
-                                    : joinQuery.shards,
-                            traceSpan: span,
-                        })) as Record<string, unknown>[]
-                    }
-                    const joinedPortion = await traceSpan.traceChildOperation(
-                        `${mainCollection.name}.query.join`,
-                        fetcher,
-                    )
-                    for (const joinedRecord of joinedPortion) {
-                        this.joinRecordToMain(shardPlan, joinedRecord)
-                    }
-                    joinedRecords = joinedRecords.concat(joinedPortion)
+                for (const joinedRecord of joinedPortion) {
+                    this.joinRecordToMain(joinPlan, joinedRecord)
                 }
-                if (Date.now() > timeLimit) {
-                    break
-                }
+                joinedRecords = joinedRecords.concat(joinedPortion)
+            }
+            if (Date.now() > timeLimit) {
+                break
             }
         }
         await QJoinQuery.fetchJoinedRecords(
@@ -342,7 +276,6 @@ export class QJoinQuery {
     }
 }
 
-type Shard = string | null
 type OnValue = string
 type MainRecord = Record<string, unknown>
-type JoinPlan = Map<Shard, Map<OnValue, MainRecord[]>>
+type JoinPlan = Map<OnValue, MainRecord[]>